+++ qt_ui/windows/basemenu/ground_forces/QGroundForcesStrategy.py
@@ -1,6 +1,8 @@
 from collections import Callable
+from datetime import timedelta
 
-from PySide2.QtWidgets import QGroupBox, QLabel, QVBoxLayout, QPushButton
+from PySide2.QtWidgets import QGroupBox, QLabel, QVBoxLayout, QPushButton, QTimeEdit
+from PySide2.QtCore import QTime
 
 from game import Game
 from game.theater import ControlPoint
@@ -31,6 +33,18 @@ class QGroundForcesStrategy(QGroupBox):
             if not enemy_cp.captured:
                 layout.addWidget(QLabel(enemy_cp.name))
                 layout.addWidget(QGroundForcesStrategySelector(self.cp, enemy_cp))
+
+                layout.addWidget(QLabel("Delay movement until"))
+                self.delayed_start_spinner = QTimeEdit()
+                self.delayed_start_spinner.setMinimumTime(QTime(0, 0))
+                self.delayed_start_spinner.setDisplayFormat("T+hh:mm:ss")
+                self.delayed_start_spinner.timeChanged.connect(self.save_delayed_start)
+                self.delayed_start_spinner.setToolTip(
+                    "Forces will not move from Waypoint 1 until the specified time has elapsed."
+                )
+                self.delayed_start_spinner.setStyleSheet("background: #1D2731;")
+                layout.addWidget(self.delayed_start_spinner)
+
                 if self.game.settings.enable_frontline_cheats:
                     advance_button = QPushButton("CHEAT: Advance")
                     advance_button.setProperty("style", "btn-danger")
@@ -59,3 +73,8 @@ class QGroundForcesStrategy(QGroupBox):
         self.game.reset_ato()
         self.game.initialize_turn()
         GameUpdateSignal.get_instance().updateGame(self.game)
+
+    def save_delayed_start(self) -> None:
+        time = self.delayed_start_spinner.time()
+        seconds = time.hour() * 3600 + time.minute() * 60 + time.second()
+        self.cp.delayed_start = timedelta(seconds=seconds)